+++ src/core/layouts/Layout.tsx
@@ -10,7 +10,6 @@ const Layout: BlitzLayout<{ title?: string; children?: React.ReactNode }> = ({
     <>
       <Head>
         <title>{title || "cmdb"}</title>
-        <link rel="icon" href="/favicon.ico" />
       </Head>
 
       {children}